Instrument MediaSearch analytics using modern Event Platform
* Adds an eventLogger plugin that gives all components a $log() method
  that can be used to log events; this is basically a wrapper around
  mw.eventLog.submit(); from components. Components can call:
      this.log( {
          action: 'some_action',
          data: 'some_data',
      } )
  The logger plugin will silently add things like a persistent
  session_id (generated using mw.user.generateRandomSessionId),
  as well as the language code for the interface language the user
  has specified (the first item in the fallback chain).
* The plugin takes 2 configuration options for now: stream name and
  schema to be used; these will be automatically included when the event
  is submitted
* Events are logged using the draft schema described here:
  https://phabricator.wikimedia.org/T263875

== Development Setup ==

To develop or test out this patch locally, you need to perform the
following steps.

1. Clone the "secondary" schemas repository somewhere that your local
environment can access it. You will also need to check out this patch:

    https://gerrit.wikimedia.org/r/c/schemas/event/secondary/+/630630

(once the new MediaSearch schema is merged, this step won't be necessary
and the schema can be accessed at https://schema.wikimedia.org/. )

2. Make sure you have the EventLogging and EventStreamConfig extensions
installed and enabled in your MediaWiki development environment.

3. Navigate to the EventLogging directory and install the dependencies
with "npm install". You'll need a relatively recent version of node, and
Mac OS users may need to set some compiler flags first; see the README
in EventLogging/devserver for more information. This README also contains
some info about configuring your local MediaWiki environment to work with
the dev server and schema files.

4. Go to extensions/EventLogging/devserver/eventgate.config.yaml and add
an appropriate path to the schema_base_uris configuration variable. This
should be a path to the on-disk location of the MediaSearch
schema files; you could also serve the files from a local web server and
provide the localhost url.

    schema_base_uris: [
      https://schema.wikimedia.org/repositories/primary/jsonschema,
      https://schema.wikimedia.org/repositories/secondary/jsonschema,
      ../../schemas/secondary/jsonschema
      # http://localhost:XXXX/secondary/jsonschema would also work
      # if you served the files from a local web server
    ]

5. Start the eventgate dev server by running:
"npm run eventgate-devserver" inside the EventLogging directory.

6. Configure your local MediaWiki environment:

    # LocalSettings.php

    # Enable EventLogging and EventStreamConfig
    wfLoadExtension( 'EventStreamConfig' );
    wfLoadExtension( 'EventLogging' );

    # Tell MediaWiki to use the local dev server:
    $wgEventLoggingServiceUri = 'http://localhost:8192/v1/events';

    # You'll also want to set the following config variables:
    $wgEventLoggingStreamNames = false;
    $wgEventLoggingQueueLingerSeconds = 1;

7. Send a test event to the dev server in the browser console like so:

    mw.eventLog.submit( 'stream_name', {
      $schema: '/analytics/media_search/1.0.0',
      session_id: "1",
      action: 'search_new'
    } )

This event should show up in the dev server console output and should
also be logged to the events.json file in the EventLogging directory.

NOTE: A different set of configuration values will need to be submitted
to mw-config for use in production (including the final stream name that
media search events will be collected under).

Bug: T258183
Depends-On: I3caaed1721a25b3c593e36c5259e1088056eb737
Change-Id: I17d910f83fa6d8183ecb7b135d1282331c58f7ef