Add copyable filename to QuickView and CopyTextLayout component
This patch adds a new base component, CopyTextLayout, which is
similar to mw.widgets.CopyTextLayout. It displays text with a
button that allows the user to copy the text to their clipboard.
After the user attempts to copy the text, a success or failure
message will be shown.

This component is used in the QuickView panel with the item's
filename and a WikiText snippet for displaying the file.

Finally, this patch includes a fix for icon-only buttons, which
I considered using as an alternate layout for the copy to
clipboard button but abandoned since we don't need it right now

Bug: T261699
Change-Id: I4fdb65c4ab7219504b6f2a8ab8bc2408e9912516