Improve MediaSearch experience in IE
IE doesn't support a couple of things that we're relying on:
- HTMLImageElement.decode()
- IntersectionObserver

This patch adds some conditionals to avoid these and get the JS
interface working in IE:
- Instead of relying on decode() to show images that were previously
painted in the PHP UI in the JS UI when they're ready, just show
the images as soon as their src is set.
- For the Image base component, if the IntersectionObserver isn't
supported, just add the src and don't lazy-load.
- For infinite scroll, if the IntersectionObserver isn't supported,
just show the "Load more" button
- For other elements that use the IO (SearchFilters and Tabs, which
scroll horizontally on small screens and use the IO to show/hide a
gradient that indicates that the user can scroll), just disable
this feature and never show the gradient.

Bug: T264136
Change-Id: If82d15a461aa14d85d2867af8f4d5b643d7d539c