Rewrite thumbnail urls to serve more common widths
We request thumbnails at a certain height rather than a certain
width, because that's how they fit in the grid. The width will
then depend on the file's dimensions.

Thumbnails are served on-demand (or pre-generated) via their
width. Unless we have a common width, there likely is no cached
version and it takes some time to generate the thumbnail (and we
end up being throttled if we request enough of those)

Because of our grid, we need to be flexible in widths, but we can
change exactly what width we request a thumbnail at, so that
we're much more likely to fetch a pre-existing thumb. They'll
be (slighly) downscaled on the client and there will be no
difference, except that they will not have taken as long to
generate.

Bug: T246792
Change-Id: Ic50848fdb77c5f5f14cfc58adb32f9e7e787daac