Thumbnails should have grid height
We receive smaller-than-requested dimensions if the source
image is smaller than what was requested. As a result,
these small images (e.g. icons) don't fit in our grid all
to well.
This rescales their dimensions to what we like best. Sure,
it stretches them a little, but other image search engine
result pages do the same thing already: it's not unexpected,
and it makes the whole thing visually more pleasing.

Change-Id: I4395782e6c4dced4fbcceb2fbf4af8501f9fe1b5