Cache the relevant entities used in media search query
This caches the matching wikibase entities for a day, so
that repeat requests (e.g. next batch of results for same
search term) can be fetched from cache rather than having
to execute the same search again.

This should decrease some load on elastic, and increase
the speed of noninitial searches.

Bug: T263238
Change-Id: I10db9a2c1b3bf354986308f5109836e67cb792c0