Test entity traversal
This allows testing traversal of related entities to
increase recall of media files using any of those entities.

Adding `&mediasearch_experimental` to the querystring will
invoke this experimental search query builder, which acts
the same as the current media search builder, but adds
entity traversal.
Eventually, because of ES constraints, only a subset of all
related entities might end up being used to find results.

Adding `&traversalProperties[]=P1&traversalProperties[]=P2`
will cause the P1 and P2 properties to be traversed; any
(valid) P-id can be used to test traversing that property.
If this param is not set, it'll default to traversing P31,
P171, P180 and P279.

Adding `&dumpQuery` will output the Wikidata query used
to find related entities.

Adding `&dumpEntities` will output all found child entities.

Adding `&dumpExtraFiles` will output a (few) link(s) to WCQS
that will let you find files that match those child entities.
This is done to work around the limitation described earlier,
which allows one to inspect *all* files that this
traversal *could* include (which, for technical issues, it
can't currently)
Note that these links are batched (because of WDCS technical
reasons), and the results are not ranked as they would in
an actual end-user search. It's simply a (few batches of)
unranked files that can be discovered via child entities.

Bug: T258055
Change-Id: Ief1926c70531013741f8f254f82e17b84972d217