Improve QuickView image loading experience
When the quickview panel is opened, figure out the height of the
image that will eventually load and use that height to style a
wrapper element around the image. When the image loads, there
won't be a layout change. On image load, remove that hard-coded
height style in case the user decides to change the viewport size.

Show a spinner during the image loading process (but delay it for
half a second to avoid showing it to users who are already
experiencing a pretty fast load time).

Finally, avoid stretching images beyond their original dimensions
in the QuickView panel, and add some padding to very small images
so the QuickView content and controls don't overlap.

Bug: T265696
Change-Id: Iaff9f96c9401f5bd358e865e933cc076612ccf4c