Adjust normalization
After scoring a much bigger batch of search terms, it looks
like the existing implementation doesn't hold up, and the
dataset it was based on wass too small to be representative.

Meanwhile also swapped the logic - instead of normalizing
the statement scores up to <token length>, we'll normalize
the others down to 1.

Bug: T262522
Change-Id: I01ead2f0a53a4f59b46ebefd409d8c818176dcee