Support multi-term search queries
With support for complex queries, we can no longer assume that
1 search query = 1 term. There may now be multiple terms, e.g.
"cat AND dog" or "bird -pig".

This lead to a significant refactor and code reshuffle where I'm simply
adding another layer in-between.

MediaQueryBuilder (renamed to MediaSearchQueryBuilder) is still the main
access point, but most of the code has been moved to the new
MediaSearchASTQueryBuilder, which will traverse the entire search query
(via AST parsed in Cirrus) & do essentially the same thing we've already
been doing, but now per clause (instead of 1 for the entire query string)

Along with simply moving code, some chunks ended up being refactored
because they were making the move complicated. Here's what:

1. The text-based matches are no longer repeated individually, but are
  iterated on in groups (language aware & language agnostic)
  They were all doing the same thing, but there were a few minor
  differences that made looping them like this a little awkward:
  - 'suggests' has no '.plain' version (I simply ignored that and still
    add it to the query regardless; since it doesn't exist, it won't
    affect the score anyway)
  - since 'suggest' only has 1 field, the field boost was different (^1
    vs ^3 & ^1); I simply altered the match weight (in
    MediaSearchProfiles) to the inverse score difference to even it out
  - "descriptions" field and the boost key in MediaSearchProfiles
    ("caption") didn't match, so I changed that...

2. I dropped the non-file-namespace boost:
  - it doesn't make sense for normal (NS_FILE-only) use of this builder
  - we never really bothered to check how much sense this boost makes in
    the first place

3. Construction of the fulltext builder() has moved from a static method
   inside the class into MediaSearchProfiles.php. Since we now have
   multiple layers (main builder, AST visitor, entities extractor,
   fetcher), it became really clunky to construct inside a class,
   forcing it to deal with a lot of data that's only relevant for its
   inside component. And making it annoying when you want to swap out
   one of those layers (as with the experimental entities traversal)

While the structure of, and some of the details about how the elastic
query is built have changed a bit, it all just still works the same and
will contribute to the same scores as it did before.
Except when there are multiple terms (via AND, OR, NOT), in which case
they'll now be evaluated individually (whereas, previously, they'd fall
back to another query handler, one that doesn't deal with structured data)

Bug: T252692
Depends-On: Ib99a84911f379fa26c311be754453f9bf2706b67
Change-Id: I7031fd44293900963133cc896a228178c6b3aeae