Use entityterms API instead of pageterms
We want the terms of the MediaInfo entity *on* the page
for which we're getting search results, which is what
entityterms is now specifically designed to do.

pageterms, OTOH, returns data for entities *linked to*
the page (and before entitytermss existed, it used to
do both) - but that's not what we need.

See https://lists.wikimedia.org/pipermail/wikidata/2020-July/014145.html

Bug: T267840
Change-Id: I48e6d72fba4b7684e71d7188f49d3b1484512cac