Improve image grid and fix QuickView thumbnail size
This patch introduces several CSS changes to improve aspects of
the results grid on the Images tab:
- Sets max-width and max-height of img element to the dimensions of the
  original image to avoid stretching very small images
- Improves display of such images by centering them within the wrapper
  box and adding some left/right padding to the wrapper to mimic the
  padding-like effect above/below the image
- Constrains initial width of wrapper for images with a large aspect
  ratio (super wide) to a limit, since these images often cause their
  neighbors to stretch too much if the calculated width is used
- Aligns result items to flex-start. Space-between doesn't work when
  there are only a couple of results.
- On larger screens, sets flex-grow to 0 for portrait-oriented images.
  These images look blurry when stretched, so not letting them grow
  removes most instances of blurriness in grid images.

Note that under some circumstances (e.g. viewport width ~800 and search
results containing lots of panoramas), there will be a gap at the end
of the occasional row. This is unavoidable when telling some images
not to grow, but all of the settings in this patch should mean that
this rarely happens.

Finally, this patch resolves an issue with the width of the QuickView
thumbnail overflowing its container.

Bug: T264034
Bug: T267978
Change-Id: I874b45f74413ff77a2fe08b4cfcb2edac46a4552