Only show tab focus styles on keyboard nav
Instead of showing a visual focus style anytime a tab is selected,
use tabindex attributes and CSS to enable us to show the visual
focus style when the user navigates via keyboard but not via mouse.

This patch also:
- Standardizes some CSS class names to match the rest of the
WBMI base components library. .is-active and .is-disabled were
changed to &--current (to avoid confusion with :active) and &--disabled
- Change non-active tab label color to match OOUI component
- Add :active style (during mousedown) to match OOUI component
- Rearrange tabs LESS so we can remove stylelint-ignore comments
- Ensures pointer cursor on hover in server-rendered UI

Bug: T265048
Change-Id: I760ed86804dde4d1e83cd0654a5e25d613ef9bfd