Make match_explorer query optional
match_explorer is a feature from the
o19s/elasticsearch-learning-to-rank elastic plugin.
While it is present on production & vagrant setups,
it might not be in other environments.

We have devs on a Docker environment where this plugin
is not installed on their elastic instance, in which
case reliance on this feature prevents media search
from working altogether.

Depends-On: I7bc6bf5dca6af46c2094b6e209c9e5d22869a2c5
Change-Id: I579a6dd5eca698f50d7a80e2190f3add94ef9fd2