Support dates with a space before time in Safari
Safari does not support dates in the format YYYY-MM-DD HH:mm:ss.
Since this is a common value we get back from extended metadata,
we should trim off the time before checking if it's a valid date,
since Safari does support YYYY-MM-DD. If it's not a valid date
(e.g. "circa 1900"), we'll just use the original value anyway.

Bug: T265078
Change-Id: I21e5e5ff7e13e8f492df8304a3404265a3f220d4