Fix normalization factor
The normalization correction should not exceed 1.25, but the
old code had it start from 1.25.

This wa imply a case of me confusing "1.25 is the maximum" with
"max", when in fact it should be "min". Happened a 100 times
before, and I'm sure I'll mix them up 100 times more in the future.

Bug: T262522
Change-Id: Id99078fa16063b1c9574409317b9c6b4a30a16d5