Add "End of Results" message
Adds an EndOfResults component and related styles; also adds equivalent
markup to the SERPWidget template.

Treating this as a new stand-alone component instead of re-using the
existing empty state one because we may introduce concept chips or
other search prompts here at a later time. For now static text is
all that is required.

Bug: T261166
Change-Id: I6943fe34b68a44c3fc975a1a4a55a3f6515d45a7