Move loading icon and loading state components into SearchResults
The loading icon should be centered under the search results grid at
all times, including when the QuickView panel is open, reducing the
width of the search results grid. Since the App component doesn't know
when that's happening, this patch moves the loading icon into the
SearchResults component so it can live in a div with the search
results list and, therefore, be the same width as that list at all times.

With this in place, the "load more" button being centered in the App
component looked weird, so I moved it to the SearchResults component
as well so that it'll line up with the loading indicator.

Also moves the various empty state components into SearchResults to
simplify the App component somewhat.

Finally, this patch adds a dynamic class for the new list wrapper div
that's present when QuickView is open, but it also maintains the
dynamic class with the same purpose on the list div itself. This
keeps the LESS code significantly simpler.

Bug: T259848
Change-Id: I6cd1799b71a9b61d5228e90cf90efbc2ff3e2423