Get CopyTextLayout working in earlier iOS versions
The previous implementation, which was based on the CopyTextLayout
widget in MediaWiki core, didn't work in iOS versions below 14.
The implementation has been updated to work in those versions, plus
all other modern browsers. It has limitations (especially the selection
range limit, which is very high but could also not be high enough
depending on the use case), but is sufficient for the MediaSearch
UI. We should rethink this if we're going to upstream this into
WVUI, though.

Bug: T261699
Change-Id: Iff0c1ea62c3ca07040cf609df163248d77e293a0