Distinguish between null continue value and unknown one
When we get an API response with search result data, if there are
no more results, there will be no continue param and the continue
value for that media type will be set to null. This must be
differentiated from the scenario where we don't yet know if there
are more results because we haven't performed a search yet.

In a recent change, the code that resets state when a new search
is performed was setting continue values to null, indicating no
more results, rather than undefined, indicating an unknown
value. This patch resets that code, and sets the initial value for
continue to an empty string rather than 0 for consistency's sake.

This patch also renames the hasMore getter to the more precise
checkForMore, and changes the logic for showing the "load more"
button to only work if there are definitely more results.

Finally, this changes logic on filter change to only clear results
for the current media type, not all types.

Bug: T272548
Change-Id: I183d20ca229932742a8aa9ab1d4af3d3e0f6b6e9