Catch formatting errors and show something useful instead
Invalid values are validated on submission, so this shouldn't
ever be called unless we mess up in code or config.
But judging from the linked bug, it seems that this has in
fact happened at some point, so... let's not crash completely.

This will simply show "Invalid value" when anything fails to
render. This will also allow the UI to load and make it
possible to remove the invalid value.

Bug: T272508
Change-Id: I4162d8a53dedb13246380b68738425e13fbd1851