Add formatted sizes and wordcount to results, and format lots of numbers
The main goal of this task was do to the following, accomplished in
this patch:
- Add formatted page size and wordcount to Page results
- Add formatted image size to Other results

This involved:
- Adding methods to the searchResult mixin that mimic Language class
methods to format file sizes and numbers. The file size formatter is
adapted from core code for the Special:Upload page. It can mostly
match the output of Language::formatSize, except it won't change out
the decimal separator per-language. Also, there's another popular
solution to generating the size string that involves a bunch of
JavaScript Math functions, but I think the solution used here is
easier to read.
- Using the searchResult mixin in the PageResult and OtherResult
components. Now all result components use this mixin.

Some additional work was done in this patch:
- Added resolution to other results in the PHP UI, hardcoded the dir
of the resolution to always be ltr in both PHP and JS, and formatted
the numbers in PHP and JS
- Formatted resolution numbers in QuickView for images
- Formatted PageResult category numbers (PHP and JS)
- Formatted total hits (PHP and JS)

Bug: T262992
Change-Id: Ief86e3a007e127547460fab859a3fd203beb7406