Instrument QuickView copy events
Updates the QuickView to log when the user copies the filename or wikitext link
using the built-in tool. Only successful copy events are logged. The
CopyTextWidget itself is considered a base component, so it should remain
agnostic of plugins and larger application structure. Instead it simply emits a
"copy" event when text is successfully copied to the user's clipboard. The
Quickview is responsible for differentiating between the two different cases
and relies in the logger plugin.

bug: T263663
depends-on: https://gerrit.wikimedia.org/r/q/If3bebf149fe3cd095cfa192f7405dd32c9ae0ca6
Change-Id: I2563af98780396c10048c43bbb3859df93fa7ba5