Avoid some cases of stretched images in the results grid
Because we're using a CSS-only solution, there are some shortcomings to
our grid, most notably that images get stretched too wide in certain
circumstances, leading to a blurry image. Since we can't use JavaScript
to find out information like how many rows there are, how many images
are on a row, or how big each image is at any given time, we'll need to
handle specific cases to make improvements. This patch does the
following:

1. One of the worst cases of stretched/blurry images is when there are
1-3 image results. There's already code to prevent the last image in
the results set from growing on large screens, but in a small results
set, the first or second image might be stretched too wide. This has
the unfortunate side effect of causing the stretched image's display
to shift drastically when the QuickView panel is opened. To fix this,
this patch adds a class to image results when there are 3 or fewer.
Styles for this class will prevent these images from growing at all,
and will limit them to 33% width. This way, the images will all be on
the same row, and none of them will be stretched wider than their
natural width (given the height of each row).

Beyond 3 image results, it's likely that there will be more than 1 row,
preventing the worst version of this issue. Still, it's not a perfect
solution, and results sets with >3 images might have an
overly-stretched second-to-last image. It's at least less noticeable
when it's not happening in the first or second row. Why not just set a
max-width on each result? Because this will cause gaps at the end of
grid rows.

2. The style that prevented the last image result from growing now
applies at all screen sizes, and whether or not the QuickView panel is
opened. This covers the case when the final image in a results set is
portrait-oriented and may appear on its own row when QuickView is open.
We don't want the image in this case to grow at all. Since it's the
final image on a row, we can allow the extra space at the end of the
row that may occur.

3. A wrapper div was added to the SearchResults component in a previous
patch that was only being used for QuickView-related styles, so it
wasn't added to the PHP UI. However, not having it there causes some
layout differences when there are only a few results. So, this patch
adds that div to the PHP UI template.

Bug: T270735
Change-Id: I56dba670990ccdfe3f7f170c949127d55d79ebfc