Add "did you mean" feature to Media Search
Displays a snippet of text with an alternate search term suggestion
if the user enters a term that seems like a typo. Suggestions are
returned when the additional "gsrinfo" param "suggestion" is
provided.

This patch adds the appropriate parameters to the places where API
requests get made (in SpecialMediaSearch.php and in the Vuex search
action). To display the alternate search term, the SERP widget
(php-rendered) has been updated. On the JS side, a new DidYouMean.vue
component has been added.

Bug: T260292
Change-Id: I233587cc6b93042b4ccbda4b29f8018613799a89