Debounce thumbnail loading
Thumbnails start to get loaded once they're inside or near
the viewport. This is perfect under normal conditions.

However, with the amount of thumbnails we might have on-page,
this might mean an absurd amount of thumbnails. If they're
meant to be displayed on screen, that's fine, but if they're
quickly being scrolled past, we can skip those.

This adds a 250ms delay between "image first detected to be
within or near viewport" and when we actually start to load
it. Under normal viewing conditions, 250ms is plenty for
images to start pre-loading off-screen ahead of time while
casually scrolling through the results.
250ms is even acceptable if the image suddenly appears within
viewport (e.g. after a rapid scroll, where it didn't have the
chance to start preloading off-screen) - 250ms is roughly
equal to the average human reaction time, so it's a barely
perceptible delay.

The benefit here is when users are scrolling through the
results really fast. In such case, this will now no longer
load images that, 250ms seconds later, are off-screen again.
This reduces both the load on the thumbnailing server, and
the local browser thumbnail request queue.

Bug: T263705
Change-Id: I52dc7fc27d9e9b8142d14e166a1dd6a532c7f195