Refactor field iteration stuff
This extracts the logic to iterate the fields from the
node handler classes because it wasn't working:
- some node handlers may not need to iterate any nodes
  at all (though those could already simply implement
  ParsedNodeHandlerInterface rather than
  AbstractTextNodeHandler)
- other node handlers may need to iterate fields more
  than once, sometimes even with different query
  clauses for each iteration
- and some might want to combine the field queries in
  a different way than what AbstractTextNodeHandler was
  currently doing (e.g. dismax functions instead of
  boolquery should clauses)

This, I hope makes things simpler: FieldIterator is now
a standalone object that you can then iterate as an array,
so it's easy to initialize multiple (different) iterators.
All it needs is a custom FieldQueryBuilderInterface
instance, that builds a specific query per field.

That specific thing is probably the least straightforward
part of this patch. I could simply allow a callback
function as well, but I decided to go with an interface
because it's the only way to enforce a specific signature
(method parameters)

Change-Id: I033cf766d9457995e003770e9746771df64bfe0e