Add English label & aliases of top related Wikidata items to query
This patch takes the label and aliases of related Wikidata items
for a search term (with over 0.5 relevance score) and adds
the following to the search query:
- A filter query that adds a phrase MultiMatch query per synonym.
Using the phrase query type means that only files that contain the
entire phrase will be added to the results set. In manual testing,
this proved to be more effective at reducing noise than using the
default best_fields type and adding in a minimum_should_match
factor.
- Ranking queries similar to the ones added for the search term,
but with a (decayed by relevance score) boost for each field to
ensure the actual search term influences the overall score the most.
Instead of evenly adding to the score for each synonym, a DisMax
query will choose the highest score out of the term itself and all
the synonyms and add it to the overall score, along with decayed
scores vor the rest of the matches. This boosts the position of
files containing synonyms within the results set, which should,
ideally, lead to the perception of more relevant results.

Finally, this patch updates the entity fetcher to use search as a
generator so we can grab labels and aliases while finding related
items.

See the Phab ticket for more information and suggested test cases.

Bug: T258053
Depends-On: Ice66403b509bd5cfe9ce40d686c76c96ad7d3a2f
Change-Id: Ie1097d5ab05726b9f7cb4b603fbdb48bf2fb0da9