Programmatically generate search options in PHP
This replaces the previous JSON files which defined filter and
sort options with a new Services class, MediaSearchOptions.

This class has a single public static method,
MediaSearchOptions::getSearchOptions, which can be invoked as
a ResourceLoader callback as well as in normal Special page code.

The class includes some hard-coded data (sort, mimetypes, etc),
but license groups are now taken from a community-maintained wiki
page (specifically, this one:
https://commons.wikimedia.org/wiki/MediaWiki:Wikibasecirrus-license-mapping).

The getSearchOptions method also handles all translation of messages
(reducing what needs to be delivered to the browser) and determines
which media types should receive what set of options (removing the
need for duplicate JS and PHP logic to handle this).

The community is free to provide new license groups going forward,
but if they do so they will also need to add on-wiki pages to
define internationalized label messages that will appear in the
dropdown menu to represent the new items.

For example, if a new license group is added to the mapping page
called "foo", a page in the MediaWiki namespace needs to be added
to label it.  That page would live at:
MediaWiki:Wikibasemediainfo-special-mediasearch-filter-license-foo.

Bug: T273166
Change-Id: I1f16410eca1378bd4f4d87bf9639d007debdfede
Depends-On: Ieec0c79c30531a3ea158ddf66357186691cabcd8