Prevent searches that are already known to have no results
The offset param's fallback to 0 could mean that a new
(existing) set of results ends up being requested, when
in fact we already have all the results that we need.

A continue value of null means that there are no results
available (in which case this request shouldn't happen),
while a continue value of undefined means that we don't
know & should continue to default to 0.

Bug: T272923
Change-Id: I561ae988b5a37c6c1eec0936c59770e820745039