Omit duplicate results
It is possible that a result if at e.g. place 40 (out of
a limit of 40) in the first request, but is then the first
of the next batch (place 41).
This could happen because of minor inconsistencies in
elastic scores across multiple replicas, but also because
circumstances have changed (e.g. content changed & it
really did go up/down in the list)

Regardless of the reason, since we continue to show
existing results on the screen, we should not add results
that we've already encountered: they'll become duplicates.

Bug: T272923
Change-Id: Id3f7ff7b1dc2c124ddde726e03865d3a7db9b220