Also requet timestamp|snippet from non-page results
These are being requested for results in the pages tab
(the only place where they're used) but the template
params code expects them to always exist.
This currently causes an "undefined index" notice
on non-file tabs, where the 'timestamp' is unconditionally
used.

There's no additional cost to always fetch them
serverside (the data is already available from the
search results, it's just a matter of including them
in the result), so we might as well be consistent and
ask them for every type of request (as already happens
on the frontend), then we can safely assume it always
exists & don't need to check for it's existence when
we do want to use it.

Bug: T271174
Change-Id: Id9cb40a0ed5d87bb5d95e99c534ffa546ae2f900