Implement support for phrase query nodes
While the ways these scores end up being calculated might be different,
I got the impression that, overall, they are quite in line with words
queries (in how matches in fields score relative to one another).
Using the same weights for the same fields will likely have similar
results between phrases & words, where matches in certain fields are
considered more relevant/important than matches in other fields in the
same way. (Sadly, I don't have a wealth of real-world data to compare
this with, nor do we currently have a convenient way of objectively
comparing large sets of searches of a specific type - I simply selected
a bunch of phrase queries from
https://superset.wikimedia.org/superset/dashboard/154 and manually
compared results, but IDK how representative these are as a sample for
phrase searches in general)

The only exception is statements, whose score & weight are calculated
differently. It's boost is heavily influenced by calculating how much
of the "snippet" actually contains the search term. Unlike for word
queries, we don't have the more focused 'titlesnippet' etc; just the
generic 'snippet' (based on the contents of the 'all' field) which is
likely to contain much more content, which will reduce the percentage
of the match, which will, if anything, lead to lower boosts for
matching statements.

Still, compared to the current situation where statements are not
included at all, that's still a big win IMO.

And while there are already some optimization embedded in 'all', they
don't necessarily match up with Commons-specific observations. Matches
in titles tend to have a massively inflated impact, when (for files)
they are not all that much better than some other fields. En plus,
scoring fields in the same way makes the codebase more consistent and
predictable.

Bug: T268235
Change-Id: I95d1e816c4cf9d9e3b62c4a40b11200c6710afb2