Handle erroneous search filter params in URL
Introduces a new validateFilters method in SpecialMediaSearch.

The method takes all the user-specified filter settings (derived
from URL params) and compares them agains the complete set of
search options (all options for all filters for all media types).
Filter parameters are considered invalid if:

1) A filter has been specified which is not supported for the
  currently-active media type (even if it is supported for other
  types)
2) An allowed filter for the current mediatype is supplied with
  an un-supported value (these are also type-specific).

Should either of these conditions be met, an error is thrown early
in the course of the SpecialPage execute method. This triggers
an early bail-out using an alternate HTML template (SearchErrorWidget)
that informs the user they attempted an invalid search and provides
a link to the plain Special:MediaSearch page with no parameters.

Bug: T271387
Change-Id: Ia9cfad8c2eae518e07323d7211c74e924c08ff37