Rename URI to params to their search keyword equivalent
This simplifies code because we no longer have to
translate between 2 terms that mean the same thing.
It is also easier for users who are already familiar
with the search terms.
And it makes validating stuff etc easier and safer,
because we can simply use existing lists of supported
terms, rather having to maintain those ourselves.

Bug: T274494
Change-Id: I26910a96c66f6c0d9d8148d7e0a5eaa7658e73be