Rename 'bitmap' to 'image'
I remember picking 'bitmap' early on so it'd be in
line with the existing media type configurations.
We've long since diverged, though, with the first
tab having become bitmap+drawing, the 'other' tab
having become a combination of a whole bunch of
things, and even a tab that is not about media at
all.
I think 'image' is simpler for most people to
understand (compared to 'bitmap'), and it'd also
help avoid confusion for those who know what
'bitmap' already means in MediaWiki.

Bug: T274494
Change-Id: I7e9ab99a5667a94a76f10d6a731364ac57d79599
Depends-On: Ibb8ba09575871c9225f90b54e9fcbcd05bdb4964