Make WikibaseCirrusSearch a soft dependency
The license map parsing code used in MediaSearchOptions
makes WikibaseCirrusSearch a hard dependency.
It already is for this repo, but it will not be once we
split this off into a separate extension.

This will first check the configured whitelist of enabled
filters before it actually calls the WikibaseCirrusSearch
code - this will allow 3rd party installs to opt out of
features that require WikibaseCirrusSearch, making it a
soft dependency.

Bug: T273166
Change-Id: Ibf1de315654a9b5375ea564548d3f81f4198b45d