Remove unused restrictions
This restriction was introduced when the proof of concept
was being tested. It lived for a couple of weeks and was
rapidly relaxed to allow access to anyone.

Might as well get rid of this altogether now: this is
becoming a default feature and should not have any
restrictions ever again.

Change-Id: I792bbf6b831318859a9622316ccbdd842c5f38fc