Add namespace filter
- Expand the existing mobile-only dialog component to support
non-fullscreen use cases. This involves appending the dialog to
the end of the body tag on mount, then using an active prop to
determine whether it should be displayed. Additional styles were
added based on differences between the mobile/fullscreen and
desktop versions.
- Add a button to trigger the namespace dialog, styled to look
similar to the other filter select lists
- Build a dialog to display namespace options and handle their
data on submit
- Add the correct namespaces when searching on the client side
- Add namespace data to search options so it can be used
across the codebase

Bug: T276261
Change-Id: If853269368bbc743d6c11857b6d3d98ef117ab19