Update preference name in prep for move to separate extension
This code will soon move into Extension:MediaSearch, where
the `sdms` prefix is used. Once we do, we'll want to make
sure that we inherit the same preference settings. Given
how this preference has no history yet, I guess it makes
most sense to simply rename it the way we'll want it to be
in the end.

Change-Id: Ie16ff1ec2d879ee9558ca8ffa8cd79298ca148fa