Support namespace filters in PHP & URL Params
Exposes the namespace filter as a new URL parameter.

The key is "namespace"; possible values include the pre-defined
namespace groups ("all", "discussion", "help") as well as a
string representing a series of custom namespace IDs separated
by pipes ("1|2|3" etc). Both the group names and the custom
ID values are validated against the set of all possible values
in PHP.

The JS UI has also been updated to handle namespace filter state
provided in PHP, and changes to the namespace filter are
reflected in URL and history state.

Bug: T276262
Change-Id: I37a468bf71d8619bbe4fca2cd250198f74bd4c4f