Make JS handling of namespace data match PHP
This changes the way the namespace data is passed around on the JS
side to mimic what's happening in the URL and on the PHP end. Instead
of submitting an object of data, just submit the namespace group, or,
if a custom list of namespaces is selected, submit a pipe-separated
list. Either way, we will only be passing string data around.

This simplifies the JS, and we can handle the data in a consistent
way across the codebase (first checking if the value is one of the
namespace groups, and, if not, assuming it's a custom list).

Bug: T276262
Change-Id: I95dfdfa670528c402753ba5e5e4c6ee141b9c2c6